fix: query conceptnet graph
